Fix: Correct HomePage and ProfilePage structure and routing
The component logic for the main layout (Navbar + Outlet), originally in `src/pages/ProfilePage.jsx` and aliased as `HomePage` in `App.jsx`, has been moved to `src/pages/HomePage.jsx`. This resolves an issue where `src/pages/HomePage.jsx` was empty.

A new placeholder component has been created for `src/pages/ProfilePage.jsx` to serve as the actual user profile page.

The import for `HomePage` in `src/App.jsx` has been updated to point to the corrected `src/pages/HomePage.jsx`.

This change ensures that:
- The main application layout is correctly defined in `HomePage.jsx`.
- Navigating to `/me` will render the `ProfilePage` component within the main layout, rather than attempting to render the layout within itself.
